requirements.txt
python-telegram-bot>=20
opencv-python
langdetect
python-dotenv
